refactor: repositories and adds review management
Moves service logic to dedicated repository classes, consolidating data access. Updates seed handling, enhances query capabilities with aggregation for ratings, and introduces review management functions for better modularity.
