+++ apps/backend/src/infrastructure/infrastructure.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { BookSchema, ReviewSchema } from '@infrastructure/core';
+import { MongooseModule, getModelToken } from '@nestjs/mongoose';
+import { Model } from 'mongoose';
+import { BookSchema, ReviewSchema, BookRepository, ReviewRepository, BookMongoDocument, ReviewMongoDocument } from '@infrastructure/core';
 
 @Module({
   imports: [
@@ -9,6 +10,22 @@ import { BookSchema, ReviewSchema } from '@infrastructure/core';
       { name: 'Review', schema: ReviewSchema }
     ])
   ],
-  exports: [MongooseModule]
+  providers: [
+    {
+      provide: BookRepository,
+      useFactory: (bookModel: Model<BookMongoDocument>) => {
+        return new BookRepository(bookModel);
+      },
+      inject: [getModelToken('Book')]
+    },
+    {
+      provide: ReviewRepository,
+      useFactory: (reviewModel: Model<ReviewMongoDocument>) => {
+        return new ReviewRepository(reviewModel);
+      },
+      inject: [getModelToken('Review')]
+    }
+  ],
+  exports: [MongooseModule, BookRepository, ReviewRepository]
 })
 export class InfrastructureModule {}
